# gen/http/creatures/server/encode_decode.go
// Code generated by goa v3.4.2, DO NOT EDIT.
//
// creatures HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/martinohmann/goatops.farm/design

package server

import (
	"context"
	"net/http"
	"strconv"

	creatures "github.com/martinohmann/goatops.farm/gen/creatures"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeListResponse returns an encoder for responses returned by the
// creatures list endpoint.
func EncodeListResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res, _ := v.(*creatures.ListResult)
		enc := encoder(ctx, w)
		body := NewListResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// EncodeGetResponse returns an encoder for responses returned by the creatures
// get endpoint.
func EncodeGetResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res, _ := v.(*creatures.GetResult)
		enc := encoder(ctx, w)
		body := NewGetResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeGetRequest returns a decoder for requests sent to the creatures get
// endpoint.
func DecodeGetRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			name string

			params = mux.Vars(r)
		)
		name = params["name"]
		payload := NewGetPayload(name)

		return payload, nil
	}
}

// EncodeGetError returns an encoder for errors returned by the get creatures
// endpoint.
func EncodeGetError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		en, ok := v.(ErrorNamer)
		if !ok {
			return encodeError(ctx, w, v)
		}
		switch en.ErrorName() {
		case "not_found":
			res := v.(*goa.ServiceError)
			enc := encoder(ctx, w)
			var body interface{}
			if formatter != nil {
				body = formatter(res)
			} else {
				body = NewGetNotFoundResponseBody(res)
			}
			w.Header().Set("goa-error", res.ErrorName())
			w.WriteHeader(http.StatusNotFound)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// EncodeRandomFactsResponse returns an encoder for responses returned by the
// creatures random-facts endpoint.
func EncodeRandomFactsResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, interface{}) error {
	return func(ctx context.Context, w http.ResponseWriter, v interface{}) error {
		res, _ := v.(*creatures.RandomFactsResult)
		enc := encoder(ctx, w)
		body := NewRandomFactsResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeRandomFactsRequest returns a decoder for requests sent to the
// creatures random-facts endpoint.
func DecodeRandomFactsRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (interface{}, error) {
	return func(r *http.Request) (interface{}, error) {
		var (
			name string
			n    *int
			err  error

			params = mux.Vars(r)
		)
		name = params["name"]
		{
			nRaw := r.URL.Query().Get("n")
			if nRaw != "" {
				v, err2 := strconv.ParseInt(nRaw, 10, strconv.IntSize)
				if err2 != nil {
					err = goa.MergeErrors(err, goa.InvalidFieldTypeError("n", nRaw, "integer"))
				}
				pv := int(v)
				n = &pv
			}
		}
		if err != nil {
			return nil, err
		}
		payload := NewRandomFactsPayload(name, n)

		return payload, nil
	}
}

// EncodeRandomFactsError returns an encoder for errors returned by the
// random-facts creatures endpoint.
func EncodeRandomFactsError(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder, formatter func(err error) goahttp.Statuser) func(context.Context, http.ResponseWriter, error) error {
	encodeError := goahttp.ErrorEncoder(encoder, formatter)
	return func(ctx context.Context, w http.ResponseWriter, v error) error {
		en, ok := v.(ErrorNamer)
		if !ok {
			return encodeError(ctx, w, v)
		}
		switch en.ErrorName() {
		case "bad_request":
			res := v.(*goa.ServiceError)
			enc := encoder(ctx, w)
			var body interface{}
			if formatter != nil {
				body = formatter(res)
			} else {
				body = NewRandomFactsBadRequestResponseBody(res)
			}
			w.Header().Set("goa-error", res.ErrorName())
			w.WriteHeader(http.StatusBadRequest)
			return enc.Encode(body)
		case "not_found":
			res := v.(*goa.ServiceError)
			enc := encoder(ctx, w)
			var body interface{}
			if formatter != nil {
				body = formatter(res)
			} else {
				body = NewRandomFactsNotFoundResponseBody(res)
			}
			w.Header().Set("goa-error", res.ErrorName())
			w.WriteHeader(http.StatusNotFound)
			return enc.Encode(body)
		default:
			return encodeError(ctx, w, v)
		}
	}
}

// marshalCreaturesCreatureToCreatureResponseBody builds a value of type
// *CreatureResponseBody from a value of type *creatures.Creature.
func marshalCreaturesCreatureToCreatureResponseBody(v *creatures.Creature) *CreatureResponseBody {
	res := &CreatureResponseBody{
		Name: v.Name,
	}
	if v.Facts != nil {
		res.Facts = make([]string, len(v.Facts))
		for i, val := range v.Facts {
			res.Facts[i] = val
		}
	}

	return res
}