# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"goatops.farm","description":"Service for obtaining your daily dose of facts about goats and other creatures.","version":"1.0"},"servers":[{"url":"https://goatops.farm"},{"url":"http://localhost:8080"}],"paths":{"/":{"get":{"tags":["static"],"summary":"Download ./static/home.html","operationId":"static#/","responses":{"200":{"description":"File downloaded"}}}},"/api/v1/creatures":{"get":{"tags":["creatures"],"summary":"list creatures","operationId":"creatures#list","responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ListResponseBody"},"example":{"creatures":[{"facts":["Vel sunt optio officiis aut quidem.","Ratione labore cupiditate unde nesciunt.","Voluptatem aspernatur expedita velit aut."],"name":"Ea saepe ut sed dolorem."},{"facts":["Vel sunt optio officiis aut quidem.","Ratione labore cupiditate unde nesciunt.","Voluptatem aspernatur expedita velit aut."],"name":"Ea saepe ut sed dolorem."}]}}}}}}},"/api/v1/creatures/{name}":{"get":{"tags":["creatures"],"summary":"get creatures","operationId":"creatures#get","parameters":[{"name":"name","in":"path","description":"Name of the creature","required":true,"schema":{"type":"string","description":"Name of the creature","example":"goat"},"example":"goat"}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetResponseBody"},"example":{"creature":{"facts":["Vel sunt optio officiis aut quidem.","Ratione labore cupiditate unde nesciunt.","Voluptatem aspernatur expedita velit aut."],"name":"Ea saepe ut sed dolorem."}}}}},"404":{"description":"Not Found response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"},"example":{"id":"3F1FKVRR","message":"Value of ID must be an integer","name":"bad_request"}}}}}}},"/api/v1/creatures/{name}/random-facts":{"get":{"tags":["creatures"],"summary":"random-facts creatures","operationId":"creatures#random-facts","parameters":[{"name":"n","in":"query","description":"Number of random facts","allowEmptyValue":true,"schema":{"type":"integer","description":"Number of random facts","example":5,"format":"int64"},"example":5},{"name":"name","in":"path","description":"Name of the creature","required":true,"schema":{"type":"string","description":"Name of the creature","example":"goat"},"example":"goat"}],"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RandomFactsResponseBody"},"example":{"facts":["Quia perspiciatis repudiandae accusantium.","Est quod velit.","Et praesentium recusandae."]}}}},"400":{"description":"Bad Request response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"},"example":{"id":"3F1FKVRR","message":"Value of ID must be an integer","name":"bad_request"}}}},"404":{"description":"Not Found response.","content":{"application/vnd.goa.error":{"schema":{"$ref":"#/components/schemas/Error"},"example":{"id":"3F1FKVRR","message":"Value of ID must be an integer","name":"bad_request"}}}}}}},"/openapi.json":{"get":{"tags":["static"],"summary":"Download ./gen/http/openapi3.json","operationId":"static#/openapi.json","responses":{"200":{"description":"File downloaded"}}}},"/swagger-ui.html":{"get":{"tags":["static"],"summary":"Download ./static/swagger-ui.html","operationId":"static#/swagger-ui.html","responses":{"200":{"description":"File downloaded"}}}}},"components":{"schemas":{"Creature":{"type":"object","properties":{"facts":{"type":"array","items":{"type":"string","example":"Repellendus ut corporis voluptas reiciendis et nulla."},"description":"Facts about the creature","example":["Et temporibus ullam est.","Veritatis et.","Dignissimos similique maxime enim."]},"name":{"type":"string","description":"Name of the creature","example":"Odit autem aut."}},"example":{"facts":["Et fugiat quis ut nesciunt quia.","Voluptates libero."],"name":"Fugiat ea eveniet."},"required":["name","facts"]},"Error":{"type":"object","properties":{"fault":{"type":"boolean","description":"Is the error a server-side fault?","example":true},"id":{"type":"string","description":"ID is a unique identifier for this particular occurrence of the problem.","example":"123abc"},"message":{"type":"string","description":"Message is a human-readable explanation specific to this occurrence of the problem.","example":"parameter 'p' must be an integer"},"name":{"type":"string","description":"Name is the name of this class of errors.","example":"bad_request"},"temporary":{"type":"boolean","description":"Is the error temporary?","example":true},"timeout":{"type":"boolean","description":"Is the error a timeout?","example":false}},"example":{"id":"3F1FKVRR","message":"Value of ID must be an integer","name":"bad_request"},"required":["name","id","message","temporary","timeout","fault"]},"GetResponseBody":{"type":"object","properties":{"creature":{"$ref":"#/components/schemas/Creature"}},"example":{"creature":{"facts":["Vel sunt optio officiis aut quidem.","Ratione labore cupiditate unde nesciunt.","Voluptatem aspernatur expedita velit aut."],"name":"Ea saepe ut sed dolorem."}},"required":["creature"]},"ListResponseBody":{"type":"object","properties":{"creatures":{"type":"array","items":{"$ref":"#/components/schemas/Creature"},"description":"List of creatures","example":[{"facts":["Vel sunt optio officiis aut quidem.","Ratione labore cupiditate unde nesciunt.","Voluptatem aspernatur expedita velit aut."],"name":"Ea saepe ut sed dolorem."},{"facts":["Vel sunt optio officiis aut quidem.","Ratione labore cupiditate unde nesciunt.","Voluptatem aspernatur expedita velit aut."],"name":"Ea saepe ut sed dolorem."}]}},"example":{"creatures":[{"facts":["Vel sunt optio officiis aut quidem.","Ratione labore cupiditate unde nesciunt.","Voluptatem aspernatur expedita velit aut."],"name":"Ea saepe ut sed dolorem."},{"facts":["Vel sunt optio officiis aut quidem.","Ratione labore cupiditate unde nesciunt.","Voluptatem aspernatur expedita velit aut."],"name":"Ea saepe ut sed dolorem."}]},"required":["creatures"]},"RandomFactsResponseBody":{"type":"object","properties":{"facts":{"type":"array","items":{"type":"string","example":"Deleniti eligendi."},"description":"Random facts about the creature","example":["Molestiae itaque odit excepturi quod voluptas.","Sequi et facere et velit voluptatem.","Debitis ad.","Sint fugit unde."]}},"example":{"facts":["Et qui autem ipsa eligendi.","Eaque rerum."]},"required":["facts"]}}},"tags":[{"name":"static","description":"Static pages and site assets"},{"name":"creatures","description":"The creatures service provides you with farm creatures and facts about them."}]}